+++ db/updates/update_6.3-6.4.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `users` ADD `pvToken` VARCHAR(255) NOT NULL AFTER `password`; 
+ALTER TABLE `users` ADD `avatar` VARCHAR(255) NOT NULL AFTER `fullName`;