+++ pages/views/ingredients/privacyData.php
@@ -5,7 +5,7 @@ if(!$_POST['ingID']){
 	return;
 }
 
-require(__ROOT__.'/inc/sec.php');
+require_once(__ROOT__.'/inc/sec.php');
 require_once(__ROOT__.'/inc/config.php');
 require_once(__ROOT__.'/inc/opendb.php');
 
@@ -14,13 +14,14 @@ $ing = mysqli_fetch_array(mysqli_query($conn, "SELECT id,isPrivate FROM ingredie
 
 ?>
 <h3>Privacy</h3>
-	<hr>
-	<table width="100%" border="0">
-		<tr>
-			<td width="9%" height="31"><a href="#" rel="tipsy" title="If enabled, ingredient will automatically excluded if you choose to upload your ingredients to PV Online.">Private:</a></td>
-			<td width="91%" colspan="5"><input name="isPrivate" type="checkbox" id="isPrivate" value="1" <?php if($ing['isPrivate'] == '1'){; ?> checked="checked"  <?php } ?>/></td>
-		</tr>
-	</table>
+<div id="prvMsg"></div>
+	<hr />
+	<div class="row">
+    	<div class="col-xs-4">
+      		<a href="#" rel="tipsy" title="If enabled, ingredient will automatically excluded if you choose to upload your ingredients to PV Online. Will also kept private from other users in your local installation if you create any.">Private:</a>
+			<input name="isPrivate" type="checkbox" id="isPrivate" value="1" <?php if($ing['isPrivate'] == '1'){; ?> checked="checked"  <?php } ?>/>
+		</div>
+    </div>
 	<hr />
 	<p><input type="submit" name="save" class="btn btn-info" id="savePrivacy" value="Save" /></p>
     
@@ -37,9 +38,14 @@ $('#privacy').on('click', '[id*=savePrivacy]', function () {
 			ingID: '<?=$ing['id'];?>',
 			isPrivate: $("#isPrivate").is(':checked'),
 		},
-		dataType: 'html',
+		dataType: 'json',
 		success: function (data) {
-			$('#ingMsg').html(data);
+			if(data.success){
+				var msg = '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>' + data.success + '</div>';
+			}else{
+				var msg ='<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>' + data.error + '</div>';
+			}
+				$('#prvMsg').html(msg);
 		}
 	});
 });