+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # CHANGELOG
+### Version 6.4
+- Fix supplier details edit modal
+- Add available var list in SDS Html Editor
+- Convert get supplier price function to return json instead of html
+- Improve privacy settings in ingredients - a message to let user know that if privacy is enabled, the ingredient data will not be available to other user in the local installation (This is a place holder for a future release)
+- Added PV Online in ingredient search provider
+- Added option to import single ingredient data from pvonline
+- Introducing tokens (pvToken) for users to login to pvOnline - This will be fully migrated in the upcoming releases
+
 ### Version 6.3
 - Add phpMyAdmin for docker images
 - Add a message to notify user that SDS document generation its in preview state